refactor: /my-jobs ページの静的UI/動的UI分離
- page.tsx: Server Componentで静的UI(タイトル「仕事管理」、5タブ)を表示
- MyJobsContent.tsx: Client Componentで動的データのみ処理
- タブをLinkコンポーネントに変更（ハイドレーション前も動作）
- これによりタイトルとタブがJS読み込み前に即座に表示される

設計パターン:
- 静的UI: Server Component（即座にHTML表示）
- 動的データ: Client Component
